Added Comments to Describe Testing
Main is difficult to write automatic tests for given the need for a request object from Spark. Therefore I have written tests that can be checked through the UI and should the majority of the code that is written in Main.java. I'm sorry for the use of github to do this but as it is only comments I felt it was acceptable.